--- SourceFile/Goblin.cpp
@@ -1,4 +1,0 @@
-#include "../headerFile/Goblin.h"
-
-
-
