--- SourceFile/Item.cpp
@@ -1,7 +1,0 @@
-#include "../headerFile/Item.h"
-
-
-string Item::getName()
-{
-	return name;
-}
