--- SourceFile/Orc.cpp
@@ -1,3 +1,0 @@
-#include "../headerFile/Orc.h"
-
-
