--- SourceFile/Troll.cpp
@@ -1,3 +1,0 @@
-#include "../headerFile/Troll.h"
-
-
