+++ headerFile/BossMonster.h
@@ -6,6 +6,6 @@ using namespace std;
 class BossMonster : public Monster
 {
 public:
-	BossMonster() : Monster("BossMonster") { name += " 드래곤"; }
+	BossMonster() {}// : Monster("BossMonster") { name += " 드래곤"; }
 	BossMonster(const std::string& _name) : Monster(_name) {}
 };
