--- headerFile/MonsterData.cpp
@@ -1,10 +1,0 @@
-#include "../headerFile/MonsterData.h"
-
-const std::unordered_map<std::string, Stats> monsterDatas =
-{
-    //이름/    체력 / 공격력 
-    {"Goblin", {20, 5}},
-    {"Troll", {30, 3}},
-    {"Orc", {25, 7}},
-    {"Dragon", {40, 15}}
-};
